# lxd: launch instances from local instance instead of image