… tests/unit/lxd/test_launcher.py