- [X] Have you signed the [CLA](http://www.ubuntu.com/legal/contributors/)?

-----
## Overview

Cache a configured LXD environment as an instance instead of an image.

## Details

On the first run of an application, an instance will be launched from an image (i.e. an image from https://cloud-images.ubuntu.com). The instance is then setup. After setup, a copy of this instance is saved (or cached) as a 'base instance'. 

This is done to reduce setup time on subsequent runs. When the application requests a new instance on a subsequent run, the base instance will be copied to create the new instance. This instance is run through a small subset of the setup, which is referred to as 'warmup'.

## Previous behavior

![image](https://user-images.githubusercontent.com/60674096/205115452-774be7af-ebdc-44ab-b955-3e7f3e8b542b.png)

## New behavior

The flowchart below describes the logic inside the `launch()` call.  The flowchart has become crowded as it's evolved - it's a result of the fixed drawing size in Google Doc's embedded drawings.

![image](https://user-images.githubusercontent.com/60674096/205114903-3c2d7a1b-506b-46bc-b6b0-406090b88b99.png)

## Notes on the PR

- `craft_providers/lxd/launcher.py` was completely refactored, since the underlying mechanism was redone.  It's may be easier to review the new file instead of the diff.
- The unit tests did not need to be changed dramatically (thanks to https://github.com/canonical/craft-providers/pull/178).  However I added more tests.
- The integration tests are mostly unchanged too, aside from swapping `base image` assertions for `base instance` assertions.
- I've added a deprecation warning for the parameter `use_snapshots` in favor of `use_base_instance`.  For now, the old parameter will still work but will log a warning.

## Performance improvement

The scenarios below were measure how long the call to `lxd.launch()` takes when building a simple snap on a typical developer laptop.
| scenario | previous time | new time | speed up |
|-----|------|----------|-----|
|clean environment|01:58.415|01:24:483|34 seconds|
|base instance exists|00:24.721|00:04.886|19 seconds|
|instance exists|00:04:145|00:04.194|no change|

(CRAFT-1339)